fix: migrate from int to long
Migrates all integer types from `int` (32-bit) to `long` (64-bit).

Could I be a bit more directed with these changes? Probably. But the it's simpler and easier to migrate all `int`s to `long`s rather than debate how large a specific ID is going to be.

Closes #30

References:
  - https://github.com/octokit/webhooks.net/issues/30
  - https://github.com/octokit/octokit.net/issues/2351
  - https://github.com/octokit/octokit.net/pull/2352
  - https://docs.microsoft.com/en-us/dotnet/csharp/language-reference/builtin-types/integral-numeric-types

+++ src/Octokit.Webhooks/Events/PingEvent.cs
@@ -12,7 +12,7 @@ namespace Octokit.Webhooks.Events
         public string Zen { get; init; } = null!;
 
         [JsonPropertyName("hook_id")]
-        public int HookId { get; init; }
+        public long HookId { get; init; }
 
         [JsonPropertyName("hook")]
         public Hook Hook { get; init; } = null!;
