+++ src/Octokit.Webhooks/Models/Issue.cs
@@ -26,13 +26,13 @@
         public string HtmlUrl { get; init; } = null!;
 
         [JsonPropertyName("id")]
-        public int Id { get; init; }
+        public long Id { get; init; }
 
         [JsonPropertyName("node_id")]
         public string NodeId { get; init; } = null!;
 
         [JsonPropertyName("number")]
-        public int Number { get; init; }
+        public long Number { get; init; }
 
         [JsonPropertyName("title")]
         public string Title { get; init; } = null!;
@@ -59,7 +59,7 @@
         public Milestone? Milestone { get; init; }
 
         [JsonPropertyName("comments")]
-        public int Comments { get; init; }
+        public long Comments { get; init; }
 
         [JsonPropertyName("created_at")]
         public string CreatedAt { get; init; } = null!;
