+++ src/Octokit.Webhooks/Models/PageBuildEvent/Build.cs
@@ -22,7 +22,7 @@
         public string Commit { get; init; } = null!;
 
         [JsonPropertyName("duration")]
-        public int Duration { get; init; }
+        public long Duration { get; init; }
 
         [JsonPropertyName("created_at")]
         public string CreatedAt { get; init; } = null!;
