+++ src/Octokit.Webhooks/Models/ReleaseAsset.cs
@@ -13,7 +13,7 @@
         public string BrowserDownloadUrl { get; init; } = null!;
 
         [JsonPropertyName("id")]
-        public int Id { get; init; }
+        public long Id { get; init; }
 
         [JsonPropertyName("node_id")]
         public string NodeId { get; init; } = null!;
@@ -31,10 +31,10 @@
         public string ContentType { get; init; } = null!;
 
         [JsonPropertyName("size")]
-        public int Size { get; init; }
+        public long Size { get; init; }
 
         [JsonPropertyName("download_count")]
-        public int DownloadCount { get; init; }
+        public long DownloadCount { get; init; }
 
         [JsonPropertyName("created_at")]
         public string CreatedAt { get; init; } = null!;
